+++ empulse/metrics/metric/strategies/log_cost_strategy.py
@@ -466,7 +466,7 @@ def _build_log_cost_equation(
 def _log_cost_loss_to_latex(
     tp_benefit: sympy.Expr, tn_benefit: sympy.Expr, fp_cost: sympy.Expr, fn_cost: sympy.Expr
 ) -> str:
-    from sympy.printing.latex import latex
+    from ..common import _latex
 
     i, N = sympy.symbols('i N')  # noqa: N806
     cost_function = (1 / N) * sympy.Sum(
@@ -477,6 +477,6 @@ def _log_cost_loss_to_latex(
         if symbol != N:
             cost_function = cost_function.subs(symbol, str(symbol) + '_i')
 
-    output = latex(cost_function, mode='plain', order=None)
+    output = _latex(cost_function)
 
     return f'$\\displaystyle {output}$'